[wip] midi scummvm
