+++ sdl2-sonic-drivers/src/drivers/MIDDriver.hpp
@@ -17,6 +17,7 @@ namespace drivers
         // TODO need to be async
         void play(const std::shared_ptr<audio::MIDI> midi) const noexcept;
         void stop();
+        void pause();
         bool isPlaying();
     private:
         void processTrack(const audio::midi::MIDITrack& track, const uint16_t division) const noexcept;
